Cherry pick refactoring
Commits 39054574 and 4edc54f5.
